pyproject.toml
[project]
name = "genai-eamples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.3",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.2",
    "langchain-text-splitters>=1.0.0",
    "openai-agents>=0.6.2",
    "python-dotenv>=1.2.1",
]
